potential memory fix

+++ coppafish/omp/coefs.py
@@ -402,8 +402,8 @@ def get_pixel_colours(nbp_basic: NotebookPage, nbp_file: NotebookPage, nbp_extra
                                                    np.arange(z_index, z_index + 1)), 
                     int(tile), transform, nbp_file, nbp_basic, nbp_extract, return_in_bounds=True, 
                     bg_scale=nbp_extract.bg_scale)
-                
-        pixel_colors_t1 = pixel_colors_t1 / colour_norm_factor
+        
+        pixel_colors_t1 = pixel_colors_t1.astype(np.float32) / colour_norm_factor
         if pixel_colors_t1.shape[0] == 0:
             q.put(list(no_output))
             return
